feat(agents): split auto-setup into overview + tool cards
