+++ messages/zh.json
@@ -36,7 +36,7 @@
   "config_section_upstreams_label": "上游",
   "config_section_upstreams_desc": "提供商池与 API Key",
   "config_section_agents_label": "Agents",
-  "config_section_agents_desc": "Claude Code / Codex / OpenCode 自动配置",
+  "config_section_agents_desc": "Agents自动配置",
   "config_section_settings_label": "设置",
   "config_section_settings_desc": "配置文件、校验与更新",
   "proxy_core_title": "内核",
@@ -150,7 +150,7 @@
   "project_links_github_label": "GitHub",
   "project_links_open": "打开",
 
-  "client_setup_title": "Claude Code / Codex / OpenCode 自动配置",
+  "client_setup_title": "Agents自动配置",
   "client_setup_desc": "一键写入本机配置文件，让工具直接接入 Token Proxy。",
   "client_setup_proxy_base_url_label": "代理 Base URL",
   "client_setup_dirty_notice": "检测到未保存的代理配置。请先保存，避免写入过期的接入信息。",
